Seperate Remove Modal Component
